# Implement all-or-nothing stream ingestion