Per-stream sharding will introduce automatic stream sharding on rate-limited streams. The ingester's current line-by-line behavior may result in duplicates when a partially-accepted stream is resharded and sent to other ingesters. This PR makes the ingester either rate limit the entire stream before attempting to write lines to a chunk. The PR

1. Separates validation, storage, and wal replay into their own functions
2. Implements a flag to rate limit the whole stream or not
3. Checks rate limits for the whole stream rather than line-by-line
4. Returns an error with information on the whole stream when rate limited