@MasslessParticle 2022-09-07
Benchmarks:
```
Original:
goos: linux
goarch: amd64
pkg: github.com/grafana/loki/pkg/ingester
cpu: 11th Gen Intel(R) Core(TM) i7-1165G7 @ 2.80GHz
Benchmark_PushStream-8              9436            135428 ns/op          167872 B/op       1606 allocs/op
PASS
ok      github.com/grafana/loki/pkg/ingester    1.326s

New:
goos: linux
goarch: amd64
pkg: github.com/grafana/loki/pkg/ingester
cpu: 11th Gen Intel(R) Core(TM) i7-1165G7 @ 2.80GHz
Benchmark_PushStream-8             10000            120120 ns/op          165196 B/op       1600 allocs/op
PASS
ok      github.com/grafana/loki/pkg/ingester    1.248s



Comparison
benchmark                  old ns/op     new ns/op     delta
Benchmark_PushStream-8     135428        120120        -11.30%

benchmark                  old allocs     new allocs     delta
Benchmark_PushStream-8     1606           1600           -0.37%

benchmark                  old bytes     new bytes     delta
Benchmark_PushStream-8     167872        165196        -1.59%
```